# perf: weaken power in twiddles generation