--- a/fft/src/roots_of_unity.rs
+++ b/fft/src/roots_of_unity.rs
@@ -22,10 +22,12 @@ pub fn get_powers_of_primitive_root<F: IsFFTField>(
     };
     let up_to = match config {
         RootsConfig::Natural | RootsConfig::NaturalInversed => count,
+        // In bit reverse form we could need as many as `(1 << count.bits()) - 1` roots
         _ => count.next_power_of_two(),
     };
 
     let mut results = Vec::with_capacity(up_to);
+    // NOTE: a nice version would be using `core::iter::successors`. However, this is 10% faster.
     results.extend((0..up_to).scan(FieldElement::one(), |state, _| {
         let res = state.clone();
         *state = &(*state) * &root;